Separated spotlessConfig spec to a new file + exempted build & idea dirs from formatting
