+++ build.gradle.kts
@@ -9,7 +9,6 @@ buildscript {
     classpath(Plugins.androidGradlePlugin)
     classpath(Plugins.kotlinGradlePlugin)
     classpath(Plugins.navSafeArgsGradlePlugin)
-    classpath(Plugins.spotlessGradlePlugin)
   }
 }
 
@@ -39,33 +38,6 @@ afterEvaluate {
         // update version to have suffix of build id
         project.version = "${project.version}-build_$buildNumber"
       }
-    }
-  }
-}
-
-fun Project.configureSpotless() {
-  apply(plugin = Plugins.BuildPlugins.spotless)
-  configure<com.diffplug.gradle.spotless.SpotlessExtension> {
-    kotlin {
-      target("**/*.kt")
-      ktlint().userData(mapOf("indent_size" to "2", "continuation_indent_size" to "2"))
-      ktfmt().googleStyle()
-      licenseHeaderFile(
-        "${project.rootProject.projectDir}/license-header.txt",
-        "package|import|class|object|sealed|open|interface|abstract "
-        // It is necessary to tell spotless the top level of a file in order to apply config to it
-        // See: https://github.com/diffplug/spotless/issues/135
-        )
-    }
-    kotlinGradle {
-      target("*.gradle.kts")
-      ktlint().userData(mapOf("indent_size" to "2", "continuation_indent_size" to "2"))
-      ktfmt().googleStyle()
-    }
-    format("xml") {
-      target("**/*.xml")
-      prettier(mapOf("prettier" to "2.0.5", "@prettier/plugin-xml" to "0.13.0"))
-        .config(mapOf("parser" to "xml", "tabWidth" to 4))
     }
   }
 }
