+++ buildSrc/build.gradle.kts
@@ -2,4 +2,12 @@ import org.gradle.kotlin.dsl.`kotlin-dsl`
 
 plugins { `kotlin-dsl` }
 
-repositories { mavenCentral() }
+repositories {
+  google()
+  gradlePluginPortal()
+  mavenCentral()
+}
+
+dependencies {
+    implementation("com.diffplug.spotless:spotless-plugin-gradle:5.11.0")
+}
