+++ buildSrc/src/main/kotlin/Plugins.kt
@@ -32,12 +32,9 @@ object Plugins {
     "org.jetbrains.kotlin:kotlin-gradle-plugin:${Dependencies.Versions.Kotlin.stdlib}"
   const val navSafeArgsGradlePlugin =
     "androidx.navigation:navigation-safe-args-gradle-plugin:${Dependencies.Versions.Androidx.navigation}"
-  const val spotlessGradlePlugin =
-    "com.diffplug.spotless:spotless-plugin-gradle:${Versions.spotlessGradlePlugin}"
 
   object Versions {
     const val androidGradlePlugin = "4.2.0-beta06"
     const val buildTools = "30.0.2"
-    const val spotlessGradlePlugin = "5.11.0"
   }
 }
